Datasets for Classification and Regression

+++ knn-classification/src/knn/KnnRegressor.cpp
@@ -5,7 +5,7 @@ double KnnRegressor::getLabelKNeighbors(std::vector<std::pair<int, double>> &dis
     std::unordered_map<int, int> labels;
     double final_label = 0;
     for (int index = 0; index < k; index++) {
-        int label = dataset->Y[distances[index].first];
+        double label = dataset->Y[distances[index].first];
         final_label += label;
     }
     return final_label / k;
